Test for refactoring SemanticAnalyzer
